fix(tangle-dapp): Remove the hardcoded provider type
