+++ libs/tangle-shared-ui/src/hooks/useSubstrateExplorerUrl.ts
@@ -3,6 +3,7 @@ import {
   isPolkadotJsDashboard,
   makeExplorerUrl,
 } from '@webb-tools/api-provider-environment/transaction/utils';
+import { useActiveWallet } from '@webb-tools/api-provider-environment/hooks/useActiveWallet';
 import { useCallback } from 'react';
 import { Hash } from 'viem';
 
@@ -11,12 +12,17 @@ import { ExplorerType } from '../types';
 
 const useSubstrateExplorerUrl = () => {
   const { network } = useNetworkStore();
+  const [activeWallet] = useActiveWallet();
 
   const getExplorerUrl = useCallback(
     (
       txOrBlockHashOrAccountAddress: Hash | string,
       variant: ExplorerVariant,
-    ): string => {
+    ): string | null => {
+      if (activeWallet === undefined) {
+        return null;
+      }
+
       const explorerBaseUrl =
         network.nativeExplorerUrl ?? network.polkadotJsDashboardUrl;
 
@@ -24,17 +30,19 @@ const useSubstrateExplorerUrl = () => {
         explorerBaseUrl,
         txOrBlockHashOrAccountAddress,
         variant,
-        ExplorerType.Substrate,
+        activeWallet.platform === 'EVM'
+          ? ExplorerType.EVM
+          : ExplorerType.Substrate,
       );
     },
-    [network.nativeExplorerUrl, network.polkadotJsDashboardUrl],
+    [activeWallet, network.nativeExplorerUrl, network.polkadotJsDashboardUrl],
   );
 
   // Choose between TX or block hash depending on the network's
   // native explorer. If it is a PolkadotJS explorer, use the
   // block hash. Otherwise, use the transaction hash (for Statescan).
   const resolveExplorerUrl = useCallback(
-    (txHash: Hash, txBlockHash: Hash): string => {
+    (txHash: Hash, txBlockHash: Hash): string | null => {
       // Default to the PolkadotJS explorer, which is always defined
       // for all Tangle networks.
       const explorerBaseUrl =
